pyproject.toml
[project]
name = "foundry-copilot-hr-policy-knowledge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agent-framework>=1.0.0rc4",
    "agent-framework-azure-ai>=1.0.0rc4",
    "agent-framework-orchestrations>=1.0.0b260311",
    "azure-ai-documentintelligence>=1.0.2",
    "azure-ai-inference>=1.0.0b9",
    "azure-ai-projects>=2.0.0",
    "azure-identity>=1.25.2",
    "azure-search-documents>=11.7.0b2",
    "azure-storage-blob>=12.28.0",
    "diagrams>=0.25.1",
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "playwright>=1.58.0",
    "plotly>=6.6.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pytest-cov>=7.0.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "ruff>=0.15.5",
]
